# docs/openapi/foshol-api.yaml
...
openapi: 3.1.0
info:
  title: Foshol Doctor API
  version: 1.0.0
  description: >
    Crop-disease advisory for rice, tomato and potato. Every case, on every decision path,
    is reviewed by a human field officer before any advice reaches the farmer.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/cases:
    post:
      tags: [cases]
      operationId: submitCase
      summary: Submit a diagnosis case (role FARMER)
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [cropId, images, fieldArea, fieldAreaUnit]
              properties:
                cropId: { type: string, format: uuid }
                noteBn: { type: string, maxLength: 2000 }
                parentCaseId: { type: string, format: uuid, description: Set when resubmitting after a rejection }
                fieldArea: { type: number, description: Farmer-reported field area (required for FO dose context) }
                fieldAreaUnit:
                  type: string
                  enum: [DECIMAL, SQ_M, SQ_FT, HECTARE, ACRE]
                cropQuantity: { type: number, nullable: true }
                cropQuantityUnit:
                  type: string
                  nullable: true
                  enum: [KG, TON, PLANTS, BIGHAS_EQUIV]
                images:
                  type: array
                  minItems: 1
                  maxItems: 3
                  items: { type: string, format: binary }
                audio: { type: string, format: binary }
      responses:
        '202':
          description: Accepted for analysis
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CaseAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422':
          description: An image failed the local quality gate; nothing was stored
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/QualityGateProblem' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    get:
      tags: [cases]
      operationId: listMyCases
      summary: The authenticated farmer's case history (role FARMER)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
      responses:
        '200':
          description: Page of case history rows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageOfFarmerCaseRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/cases/{caseId}:
    get:
      tags: [cases]
      operationId: getCase
      parameters: [ { $ref: '#/components/parameters/CaseId' } ]
      responses:
        '200':
          description: Case detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CaseDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/review/tasks/{taskId}:
    get:
      tags: [review]
      operationId: getReviewTask
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      responses:
        '200':
          description: Everything the officer needs to decide
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewCaseDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/review/tasks/{taskId}/claim:
    post:
      tags: [review]
      operationId: claimReviewTask
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      responses:
        '200':
          description: Claimed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewTask' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/review/tasks/{taskId}/release:
    post:
      tags: [review]
      operationId: releaseReviewTask
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      responses:
        '200':
          description: Returned to the pool
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewTask' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/review/tasks/{taskId}/transfer:
    post:
      tags: [review]
      operationId: transferReviewTask
      summary: Move a live claim to another officer in the same district
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetOfficerId]
              properties:
                targetOfficerId: { type: string, format: uuid }
                expectedVersion: { type: integer }
      responses:
        '200':
          description: Transferred
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewTask' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/review/tasks/bulk-transfer:
    post:
      tags: [review]
      operationId: bulkTransferReviewTasks
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BulkTransferRequest' }
      responses:
        '200':
          description: Per-item results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BulkOperationResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/review/tasks/bulk-approve:
    post:
      tags: [review]
      operationId: bulkApproveReviewTasks
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BulkApproveRequest' }
      responses:
        '200':
          description: Per-item results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BulkOperationResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/review/tasks/bulk-reject:
    post:
      tags: [review]
      operationId: bulkRejectReviewTasks
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BulkRejectRequest' }
      responses:
        '200':
          description: Per-item results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BulkOperationResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/review/tasks/{taskId}/symptoms:
    post:
      tags: [review]
      operationId: recordOfficerSymptoms
      description: Officer-observed symptoms; recorded with source OFFICER.
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symptomIds]
              properties:
                symptomIds:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        '204': { description: Recorded }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/review/tasks/{taskId}/approve:
    post:
      tags: [review]
      operationId: publishAdvisory
      summary: Approve, edit or replace, and publish the advisory
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishAdvisoryRequest' }
      responses:
        '201':
          description: Advisory published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Advisory' }
          headers:
            Location: { schema: { type: string } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/review/tasks/{taskId}/reject:
    post:
      tags: [review]
      operationId: rejectCase
      description: Terminal. The farmer is invited to submit a new case linked by parentCaseId.
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RejectCaseRequest' }
      responses:
        '200':
          description: Rejected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rejection' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CaseId:    { name: caseId,    in: path, required: true, schema: { type: string, format: uuid } }
...
    TaskId:    { name: taskId,    in: path, required: true, schema: { type: string, format: uuid } }
    Page:      { name: page, in: query, required: false, schema: { type: integer, minimum: 0, default: 0 } }
    Size:      { name: size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  responses:
    BadRequest:
      description: Validation failure
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unauthorized:
      description: Missing or invalid credentials
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: Authenticated but not permitted
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: Unknown, or not visible to this caller
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: Illegal state transition, or idempotency key reused with a different body
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    PayloadTooLarge:
      description: Upload exceeds the configured limit
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    UnsupportedMediaType:
      description: Detected content type is not allowed
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    ServiceUnavailable:
      description: A required downstream is unavailable
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details (COMMON-API-002)
      required: [type, title, status, code, correlationId]
      properties:
        type:          { type: string, format: uri }
        title:         { type: string }
        status:        { type: integer }
        detail:        { type: string }
        instance:      { type: string }
        code:          { type: string, description: A constant from common.ErrorCodes }
        correlationId: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field:   { type: string }
              message: { type: string }
    QualityGateProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            rejectedImages:
              type: array
              items:
                type: object
                properties:
                  position: { type: integer }
                  reason:
                    type: string
                    enum: [BLURRY, UNDEREXPOSED, OVEREXPOSED, TOO_SMALL, UNREADABLE, NOT_A_CROP]
                  messageBn: { type: string }
...
    Remedy:
      type: object
      required: [id, diseaseId, type, titleBn, stepsBn, sourceRef]
      properties:
        id:        { type: string, format: uuid }
        diseaseId: { type: string, format: uuid }
        type:      { type: string, enum: [CULTURAL, ORGANIC, BIOLOGICAL, CHEMICAL] }
        titleBn:   { type: string }
        stepsBn:   { type: array, items: { type: string } }
        dosageBn:  { type: string, nullable: true }
        phiDays:   { type: integer, nullable: true, description: Mandatory when type is CHEMICAL }
        costTier:  { type: string, enum: [LOW, MEDIUM, HIGH] }
        efficacy:  { type: string, enum: [LOW, MEDIUM, HIGH] }
        sourceRef: { type: string }
        rateAmount: { type: number, nullable: true, description: Human-owned rate; null until content-owner C15 }
        rateUnit:   { type: string, nullable: true, enum: [ML, G, KG, L] }
        rateBasis:
          type: string
          nullable: true
          enum: [PER_DECIMAL, PER_SQ_M, PER_HECTARE, PER_ACRE, PER_SQ_FT, FIXED]
        rateNotesBn: { type: string, nullable: true }
        computedDose:
          $ref: '#/components/schemas/ComputedDose'
          nullable: true
          description: >
            Present on FO suggestedRemedies when rate columns and case field area allow a calculation.
            Absent on knowledge catalogue responses and on published advisory remedy refs without case context.
    ComputedDose:
      type: object
      required: [amount, unit, basis, fromArea, fromAreaUnit]
      properties:
        amount:       { type: number }
        unit:         { type: string, enum: [ML, G, KG, L] }
        basis:
          type: string
          enum: [PER_DECIMAL, PER_SQ_M, PER_HECTARE, PER_ACRE, PER_SQ_FT, FIXED]
        fromArea:     { type: number, nullable: true }
        fromAreaUnit: { type: string, nullable: true, enum: [DECIMAL, SQ_M, SQ_FT, HECTARE, ACRE] }
...
    CaseAccepted:
      type: object
      required: [caseId, status]
      properties:
        caseId: { type: string, format: uuid }
        status: { type: string, enum: [SUBMITTED] }
    CaseStatus:
      type: string
      enum: [SUBMITTED, ANALYSING, ANALYSED, IN_REVIEW, ADVISED, REJECTED, FAILED]
    DecisionPath:
      type: string
      enum: [PRIMARY, SECONDARY, UNDETERMINED]
    AnalysisMode:
      type: string
      enum: [REPLAY, LIVE]
    CaseImage:
      type: object
      required: [imageId, position, primary]
      properties:
        imageId:      { type: string, format: uuid }
        position:     { type: integer }
        primary:      { type: boolean }
        qualityScore: { type: number, format: float, nullable: true }
        width:        { type: integer, nullable: true }
        height:       { type: integer, nullable: true }
    CaseDetail:
      type: object
      required: [caseId, cropId, status, images, submittedAt, fieldArea, fieldAreaUnit, metricsSource]
      properties:
        caseId:       { type: string, format: uuid }
        cropId:       { type: string, format: uuid }
        cropNameBn:   { type: string }
        status:       { $ref: '#/components/schemas/CaseStatus' }
        decisionPath: { $ref: '#/components/schemas/DecisionPath' }
        noteBn:       { type: string, nullable: true }
        parentCaseId: { type: string, format: uuid, nullable: true }
        fieldArea:    { type: number }
        fieldAreaUnit:
          type: string
          enum: [DECIMAL, SQ_M, SQ_FT, HECTARE, ACRE]
        cropQuantity: { type: number, nullable: true }
        cropQuantityUnit:
          type: string
          nullable: true
          enum: [KG, TON, PLANTS, BIGHAS_EQUIV]
        metricsSource:
          type: string
          enum: [FORM, SPEECH, FORM_AND_SPEECH]
        images:       { type: array, items: { $ref: '#/components/schemas/CaseImage' } }
        audio:
          type: object
          nullable: true
          properties:
            audioId:       { type: string, format: uuid }
            durationMs:    { type: integer }
            transcriptBn:  { type: string, nullable: true }
        submittedAt:  { type: string, format: date-time }
...
    FarmerCaseRow:
      type: object
      required: [caseId, cropNameBn, status, submittedAt]
      properties:
        caseId:              { type: string, format: uuid }
        cropNameBn:          { type: string }
        status:              { $ref: '#/components/schemas/CaseStatus' }
        decisionPath:        { $ref: '#/components/schemas/DecisionPath' }
        diseaseNameBn:       { type: string, nullable: true }
        officerName:         { type: string, nullable: true }
        advisoryVersion:     { type: integer, nullable: true }
        rejectionMessageBn:  { type: string, nullable: true }
        thumbnailImageId:    { type: string, format: uuid, nullable: true }
        submittedAt:         { type: string, format: date-time }
        publishedAt:         { type: string, format: date-time, nullable: true }
    Candidate:
      type: object
      required: [diseaseId, diseaseCode, diseaseNameBn, confidence, rank, source]
      properties:
        diseaseId:     { type: string, format: uuid }
        diseaseCode:   { type: string }
        diseaseNameBn: { type: string }
        confidence:    { type: number, format: float }
        rank:          { type: integer }
        source:        { type: string, enum: [MODEL, KB, MERGED] }
    ExtractedSymptom:
      type: object
      required: [symptomId, symptomCode, nameBn, score, source]
      properties:
        symptomId:   { type: string, format: uuid }
        symptomCode: { type: string }
        nameBn:      { type: string }
        score:       { type: number, format: float }
        source:      { type: string, enum: [SPEECH, VISION, OFFICER] }
        matcher:     { type: string, enum: [VECTOR, FUZZY, MANUAL], nullable: true }
    AnalysisDetail:
      type: object
      required: [caseId, decisionPath, mode, candidates, symptoms, thresholds]
      properties:
        caseId:              { type: string, format: uuid }
        decisionPath:        { $ref: '#/components/schemas/DecisionPath' }
        mode:                { $ref: '#/components/schemas/AnalysisMode' }
        top1Confidence:      { type: number, format: float, nullable: true }
        top2Confidence:      { type: number, format: float, nullable: true }
        margin:              { type: number, format: float, nullable: true }
        candidates:          { type: array, items: { $ref: '#/components/schemas/Candidate' } }
        symptoms:            { type: array, items: { $ref: '#/components/schemas/ExtractedSymptom' } }
        transcriptBn:        { type: string, nullable: true }
        asrConfidence:       { type: number, format: float, nullable: true }
        hasGradcam:          { type: boolean }
        unmappedLabels:      { type: array, items: { type: string } }
        visionModelId:       { type: string, nullable: true }
        visionModelVersion:  { type: string, nullable: true }
        latencyMs:           { type: integer }
        errorCode:           { type: string, nullable: true }
        thresholds:          { $ref: '#/components/schemas/Thresholds' }
    Thresholds:
      type: object
      description: Read-only; the two routing thresholds the confidence bars draw.
      required: [high, low]
      properties:
        high: { type: number, format: float }
        low:  { type: number, format: float }
...
    ReviewTask:
      type: object
      required: [taskId, caseId, state, version, slaDueAt]
      properties:
        taskId:       { type: string, format: uuid }
        caseId:       { type: string, format: uuid }
        state:        { type: string, enum: [PENDING, CLAIMED, DONE, REJECTED] }
        officerId:    { type: string, format: uuid, nullable: true }
        officerName:  { type: string, nullable: true }
        claimedAt:    { type: string, format: date-time, nullable: true }
        claimExpiresAt: { type: string, format: date-time, nullable: true }
        slaDueAt:     { type: string, format: date-time }
        assignmentDueAt: { type: string, format: date-time, nullable: true }
        resolutionDueAt: { type: string, format: date-time, nullable: true }
        requeueCount: { type: integer }
        version:      { type: integer, description: Optimistic lock; echo it back on write }
    ReviewCaseDetail:
      type: object
      required: [task, case, analysis, suggestedRemedies]
      properties:
        task:              { $ref: '#/components/schemas/ReviewTask' }
        case:              { $ref: '#/components/schemas/CaseDetail' }
        farmerName:        { type: string }
        analysis:          { $ref: '#/components/schemas/AnalysisDetail' }
        suggestedDiseaseId: { type: string, format: uuid, nullable: true }
        suggestedRemedies: { type: array, items: { $ref: '#/components/schemas/Remedy' } }
        priorAdvisory:     { $ref: '#/components/schemas/Advisory' }
    PublishAdvisoryRequest:
      type: object
      required: [action, diseaseId, remedyIds, expectedVersion]
      properties:
        action:          { type: string, enum: [APPROVED, EDITED, REPLACED] }
        diseaseId:       { type: string, format: uuid }
        remedyIds:       { type: array, items: { type: string, format: uuid } }
        officerNoteBn:   { type: string, nullable: true, maxLength: 2000 }
        expectedVersion: { type: integer, description: The review task version the officer holds }
    RejectCaseRequest:
      type: object
      required: [reasonCode, messageBn, expectedVersion]
      properties:
        reasonCode:
          type: string
          enum: [BLURRY_IMAGE, NOT_A_CROP, WRONG_CROP, INSUFFICIENT_DETAIL, INAUDIBLE_AUDIO, OTHER]
        messageBn:       { type: string, maxLength: 1000 }
        expectedVersion: { type: integer }
    Advisory:
      type: object
      required: [advisoryId, caseId, officerId, officerName, action, version, remedies, publishedAt]
      properties:
        advisoryId:    { type: string, format: uuid }
        caseId:        { type: string, format: uuid }
        diseaseId:     { type: string, format: uuid, nullable: true }
        diseaseNameBn: { type: string, nullable: true }
        officerId:     { type: string, format: uuid }
        officerName:   { type: string }
        action:        { type: string, enum: [APPROVED, EDITED, REPLACED] }
        officerNoteBn: { type: string, nullable: true }
        version:       { type: integer }
        supersedesId:  { type: string, format: uuid, nullable: true }
        remedies:      { type: array, items: { $ref: '#/components/schemas/Remedy' } }
        publishedAt:   { type: string, format: date-time }
    Rejection:
      type: object
      required: [caseId, officerId, officerName, reasonCode, messageBn, createdAt]
      properties:
        caseId:      { type: string, format: uuid }
        officerId:   { type: string, format: uuid }
        officerName: { type: string }
        reasonCode:
          type: string
          enum: [BLURRY_IMAGE, NOT_A_CROP, WRONG_CROP, INSUFFICIENT_DETAIL, INAUDIBLE_AUDIO, OTHER]
        messageBn:   { type: string }
        createdAt:   { type: string, format: date-time }
...
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:          { type: integer }
        size:          { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages:    { type: integer }
    PageOfFarmerCaseRow:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content: { type: array, items: { $ref: '#/components/schemas/FarmerCaseRow' } }
...
    BulkTaskItem:
      type: object
      required: [taskId]
      properties:
        taskId: { type: string, format: uuid }
        expectedVersion: { type: integer }
    BulkTransferRequest:
      type: object
      required: [targetOfficerId, items]
      properties:
        targetOfficerId: { type: string, format: uuid }
        items: { type: array, items: { $ref: '#/components/schemas/BulkTaskItem' } }
    BulkApproveRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            allOf:
              - { $ref: '#/components/schemas/PublishAdvisoryRequest' }
              - type: object
                required: [taskId]
                properties:
                  taskId: { type: string, format: uuid }
    BulkRejectRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            allOf:
              - { $ref: '#/components/schemas/RejectCaseRequest' }
              - type: object
                required: [taskId]
                properties:
                  taskId: { type: string, format: uuid }
    BulkOperationResult:
      type: object
      required: [succeeded, failed, results]
      properties:
        succeeded: { type: integer }
        failed: { type: integer }
        results:
          type: array
          items:
            type: object
            required: [taskId, status]
            properties:
              taskId: { type: string, format: uuid }
              status: { type: string, enum: [OK, FAILED] }
              errorCode: { type: string, nullable: true }